% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{get_query}
\alias{get_query}
\title{get_query(num_chamado,url_sql, url_sheets, credspath, date_change = FALSE, sheetname = 'Pagina1')}
\usage{
get_query(
  num_chamado,
  url_sql,
  url_sheets,
  credspath,
  date_change = FALSE,
  sheetname = "Pagina1"
)
}
\arguments{
\item{num_chamado}{número do chamado dessa query.}

\item{url_sql}{URl da localização do SQL}

\item{url_sheets}{url onde será escrita a tabela.}

\item{credspath}{caminho do arquivo em json com as credenciais de acesso para a base e email, deve ser armazenado localmente.}

\item{date_change}{Parâmetro se a data-hora será convertida para o fuso correto, preferencialmente para impedir inconsistências de datas, TRUE/FALSE,default é FALSE.}

\item{sheetname}{Parâmetro com o nome da aba, default é 'Pagina1'}
}
\description{
Função para extração e escrever uma query apenas, sem filtrar
}
\examples{
num_chamado = '1111'
url_sql='https://docs.google.com/spreadsheets/TESTE'
url_sheets = 'C:/Users/Usuario/Desktop/creds.json'
get_query(num_chamado,url_sql, url_sheets, credspath, date_change = FALSE, sheetname = 'Pagina1')
}
