% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query_auto}
\alias{get_query_auto}
\title{Conecta e realiza à consulta à base, retorna a query}
\usage{
get_query_auto(usernamedb, passwordb, dbname, query)
}
\value{
Dataframe query.
}
\description{
Conecta e realiza à consulta à base, retorna a query
}
\examples{
get_query_auto('user', 'senha','nome_base','SELECT * FROM TABLE')
}
