% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{get_write_query_sqlserver}
\alias{get_write_query_sqlserver}
\title{get_write_query_sqlserver(num_chamado,url_sql, url_sheets, credspath, date_change = FALSE,sheetname="Pagina1")}
\usage{
get_write_query_sqlserver(
  num_chamado,
  url_sql,
  url_sheets,
  credspath,
  date_change = FALSE,
  sheetname = "Pagina1",
  encoding = "pt_BR.UTF-8"
)
}
\arguments{
\item{num_chamado}{número do chamado dessa query.}

\item{url_sql}{URL da localização do SQL}

\item{url_sheets}{url onde será escrita a tabela.}

\item{credspath}{caminho do arquivo em json com as credenciais de acesso para a base e email, deve ser armazenado localmente.}

\item{date_change}{Parâmetro se a data-hora será convertida para o fuso correto, preferencialmente para impedir inconsistências de datas, TRUE/FALSE,default é FALSE.}

\item{sheetname}{Parâmetro com o nome da aba, default é 'Pagina1'}

\item{encoding}{Codificação para caracteres especiais, o default é "pt_BR.UTF-8", caso deseje mudar para Latin-1 troque por "Portuguese_Brazil.1252"}
}
\description{
Função para extração e escrever uma query apenas, usando a conexão do sql server, que funcione, é necessario q o driver odbc esteja funcionante já
para instruções: https://turbofuture.com/computers/Connect-to-SQL-Server-from-R
}
