% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{gw_query_auto}
\alias{gw_query_auto}
\title{gw_query_auto()
Função para automatização total porém, com o objetivo de execução'}
\usage{
gw_query_auto(
  url_pedidos,
  credspath,
  periodo = "Diário",
  status = "ATIVO",
  date_change = FALSE
)
}
\description{
gw_query_auto()
Função para automatização total porém, com o objetivo de execução'
}
\examples{
NOT RUN
url_pedidos='https://docs.google.com/spreadsheets/TESTE'
credspath = 'C:/Users/Usuario/Desktop/creds.json'
gw_query_test(url_pedidos, credspath,periodo='Homolog',status='TESTE', date_change = TRUE)
}
