% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{gw_query_test}
\alias{gw_query_test}
\title{gw_query_test(url_pedidos, credspath,periodo='Homolog',status='TESTE', date_change = FALSE)}
\usage{
gw_query_test(
  url_pedidos,
  credspath,
  periodo = "Homolog",
  status = "TESTE",
  date_change = FALSE
)
}
\arguments{
\item{url_pedidos}{url ou link onde estão as informações para o loop.}

\item{credspath}{caminho do arquivo em json com as credenciais de acesso para a base e email, deve ser armazenado localmente.}

\item{periodo}{periodicidade das queries, deve ser selecionada/ser igual a Diario/Mensal/Homolog/Semanal.}

\item{status}{status do pedido na tabela das queries,  deve ser selecionada/ser igual ATIVO,Cancelada,TESTE}

\item{date_change}{parâmetro se a data-hora será convertida para o fuso correto, preferencialmente para impedir inconsistências de datas .}
}
\description{
Função para automatização total porém, com o objetivo de teste
}
