% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{log_error}
\alias{log_error}
\title{log_error_gen(error_msg)}
\usage{
log_error(error_msg, num_chamado)
}
\arguments{
\item{error_msg}{Mensagem de erro que será impressa caso haja erro no chamado}

\item{num_chamado}{Número do chamado passado automaticamente para a função, em caso de erro}
}
\description{
Função para erro por chamado
}
