% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{log_error_gen}
\alias{log_error_gen}
\title{log_error_gen()
Função para automatização total porém, com o objetivo de teste}
\usage{
log_error_gen(error_msg)
}
\description{
log_error_gen()
Função para automatização total porém, com o objetivo de teste
}
