% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sheets.R
\name{write_query_sheet}
\alias{write_query_sheet}
\title{write_query_sheet()
Função para escrever a tabela recuperada pela query em um documento do sheets.
#' @df dataframe correspondente à consulta realizada a partir do sql.}
\usage{
write_query_sheet(df, url_destiny, sheetname = "Página1", date_change = FALSE)
}
\description{
write_query_sheet()
Função para escrever a tabela recuperada pela query em um documento do sheets.
#' @df dataframe correspondente à consulta realizada a partir do sql.
}
