% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query_auto}
\alias{get_query_auto}
\title{Conecta e realiza à consulta à base, retorna a query}
\usage{
get_query_auto(usernamedb, passwordb, dbname, query)
}
\arguments{
\item{usernamedb}{credencial de usuário para acesso à base.}

\item{passwordb}{credencial de senha para acesso à base.}

\item{dbname}{A number.}

\item{query}{Código em sql, preferencialmente em string para realizar a consulta.}
}
\value{
Dataframe query.
}
\description{
Conecta e realiza à consulta à base, retorna a query
}
