% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query_sqlserver}
\alias{get_query_sqlserver}
\title{Conecta e realiza à consulta à base, retorna a query
Para setar a conexão com o SQL server, primeiramente necessário configurar o driver, com  a base, ver
https://turbofuture.com/computers/Connect-to-SQL-Server-from-R}
\usage{
get_query_sqlserver(query)
}
\arguments{
\item{query}{Código em sql, preferencialmente em string para realizar a consulta.}
}
\value{
Dataframe query.
}
\description{
Conecta e realiza à consulta à base, retorna a query
Para setar a conexão com o SQL server, primeiramente necessário configurar o driver, com  a base, ver
https://turbofuture.com/computers/Connect-to-SQL-Server-from-R
}
