% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{gfw_query}
\alias{gfw_query}
\title{Função para extração e escrever uma query apenas, com uma tabela de filtro, escrevendo a tabela resultante no sheets.}
\usage{
gfw_query(params)
}
\arguments{
\item{params}{lista com os parâmetros da execução da função, deve ser da seguinte forma:params <- list(num_chamado = 123,url_sql = "your_sql_url",url_sheets = "your_sheets_url",url_tabela_filtro = "url da tabela de filtro",date_change = FALSE,DTHR_MIN_PERIODO = 'Data hora minima de filtro',FILTR_COL='Nome da coluna de filtro',DTHR_MAX_PERIODO = 'Data hora máxima de filtro').}
}
\description{
Função para extração e escrever uma query apenas, com uma tabela de filtro, escrevendo a tabela resultante no sheets.
}
