% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{gw_query_auto}
\alias{gw_query_auto}
\title{gw_query_auto}
\usage{
gw_query_auto(
  url_pedidos,
  credspath,
  periodo = "Diario",
  status = "ATIVO",
  date_change = FALSE
)
}
\arguments{
\item{url_pedidos}{url ou link onde estão as informações para o loop.}

\item{credspath}{caminho do arquivo em json com as credenciais de acesso para a base e email, deve ser armazenado localmente.}

\item{periodo}{periodicidade das queries, deve ser selecionada/ser igual a Diario/Mensal/Homolog/Semanal.}

\item{status}{status do pedido na tabela das queries,  deve ser selecionada/ser igual ATIVO,Cancelada,TESTE}

\item{date_change}{Parâmetro se a data-hora será convertida para o fuso correto, preferencialmente para impedir inconsistências de datas, TRUE/FALSE.}
}
\description{
Função para automatização total porém, com o objetivo de execução
}
