% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{gw_query_test}
\alias{gw_query_test}
\title{Função para automatização total porém, com o objetivo de teste#'}
\usage{
gw_query_test(
  url_pedidos,
  credspath,
  periodo = "Homolog",
  status = "TESTE",
  date_change = FALSE
)
}
\description{
Função para automatização total porém, com o objetivo de teste#'
}
\examples{
NOT RUN
url_pedidos='https://docs.google.com/spreadsheets/TESTE'
credspath = 'C:/Users/Usuario/Desktop/creds.json'
gw_query_test(url_pedidos, credspath,periodo='Homolog',status='TESTE', date_change = TRUE)
}
