% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{log_error}
\alias{log_error}
\title{log_error()
Função para automatização total porém, com o objetivo de teste}
\usage{
log_error(error_msg, num_chamado)
}
\description{
log_error()
Função para automatização total porém, com o objetivo de teste
}
