% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{log_error_gen}
\alias{log_error_gen}
\title{log_error_gen(error_msg)}
\usage{
log_error_gen(error_msg)
}
\arguments{
\item{error_msg}{Mensagem de erro que será impressa caso haja erro geral}
}
\description{
Função para erro geral
}
