% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql.R
\name{read_sql}
\alias{read_sql}
\title{read_sql()
Função específica para ler o código em sql e  retorná-lo em formato de string}
\usage{
read_sql(email, url)
}
\value{
retorna o código em sql, em formato de string para utilização na função que faz a consulta à query
}
\description{
read_sql()
Função específica para ler o código em sql e  retorná-lo em formato de string
}
\examples{
NOT RUN
url_pedidos='https://docs.google.com/spreadsheets/TESTE'
credspath = 'C:/Users/Usuario/Desktop/creds.json'
gw_query_test(url_pedidos, credspath,periodo='Homolog',status='TESTE', date_change = TRUE)
}
