% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sheets.R
\name{write_query_sheet}
\alias{write_query_sheet}
\title{write_query_sheet()
Função para escrever a tabela recuperada pela query em um documento do sheets.
#'@param df dataframe correspondente à consulta realizada a partir do sql.}
\usage{
write_query_sheet(df, url_destiny, sheetname = "Página1", date_change = FALSE)
}
\arguments{
\item{url_destiny}{url onde será escrito a tabela.}

\item{sheetname}{nome da aba em que será escrita a tabela.}

\item{date_change}{Parâmetro que realiza a mudança de formato e fuso da data, TRUE/FALSE.}
}
\description{
write_query_sheet()
Função para escrever a tabela recuperada pela query em um documento do sheets.
#'@param df dataframe correspondente à consulta realizada a partir do sql.
}
